New SMS package for improved organization
This application's initial name for now will be SMS, which stands for
Some Maze Service (or Simple Maze Service if I change my mind). Instead
of having a mazes package as root with an ever growing number of Maze
files, it's better to have a mazes package as part of a more generic
package containing the application.