Tested and implemented missed grid method
Grid method "displayAllCells" we not implemented as the default Grid
behavior caused the existing tests to pass without being correct. This
missed implementation would have caused issues if a grid is configured
to show a path and then this method is called. It would not have worked.
Coverred issue with tests and implemented method.