Maze creation and solving animations
All maze algorithms can now be observed. First usage of this feature is
the GridAnimator class, which can be used to create an animation of a
grid as it is being manipulated by a maze-creation/maze-solving
algorithm. Created new CellGroup and CellPath data classes for
abstracting out these common concepts.